IDV-716 When using downsampled signals with processing, the appropriate (*) are now added correctly. To achieve this, the following improvements have been implemented:
- For the "Variable" column: During processing, the child signals are checked. If all of them are downsampled, the resulting signal is also marked as downsampled. However, if at least one of them is not downsampled, the resulting signal is treated as raw data.

- For the "x, y, z" columns: The same logic is applied, evaluating whether the signals they depend on are downsampled or not.

Additionally, a critical change was introduced to preserve the original data of the dependent signals. Previously, when align() was called, the signals data was modified directly, which led to the loss of the original datasets. This has now been resolved by using copy.deepcopy() when working with dependent signals, ensuring their original data remains intact.